python_backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-multipart
bcrypt
pyjwt
python-dotenv
openai
requests
passlib
